# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fitness Tracking Platform API
  version: 1.0.0
  description: Authoritative REST API for multi-user fitness management, daily tracking, workout/diet plan builders, progress analysis, reminders, and administration.
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
paths:
...
  /me/weight:
    get:
      summary: Get current weight and target
      tags: [Weight]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        '200':
          description: Weight status
    post:
      summary: Log body weight
      tags: [Weight]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weightKg]
              properties:
                weightKg: { type: number, minimum: 20, maximum: 500 }
                date: { type: string, format: date }
                notes: { type: string }
                clientOperationId: { type: string }
      responses:
        '200':
          description: Logged weight entry
    put:
      summary: Update/upsert body weight entry
      tags: [Weight]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weightKg]
              properties:
                weightKg: { type: number, minimum: 20, maximum: 500 }
                date: { type: string, format: date }
                notes: { type: string }
                clientOperationId: { type: string }
      responses:
        '200':
          description: Logged weight entry
  /me/weight/history:
    get:
      summary: Get weight logging history and rolling averages
      tags: [Weight]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        '200':
          description: Weight history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      currentWeightKg: { type: number, nullable: true }
                      rolling7DayAvg: { type: number, nullable: true }
                      rollingAverageKg: { type: number, nullable: true }
                      goal: { type: object, nullable: true }
                      entries:
                        type: array
                        items: { type: object }
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /me/water:
    post:
      summary: Log water intake
      tags: [Water]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountMl]
              properties:
                amountMl: { type: integer, minimum: 1, maximum: 10000 }
                date: { type: string, format: date }
                clientOperationId: { type: string }
      responses:
        '201':
          description: Water logged
  /me/water/today:
    get:
      summary: Get today water intake summary and entries
      tags: [Water]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Today water summary
  /me/water/history:
    get:
      summary: Get historical water logs
      tags: [Water]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 60 }
      responses:
        '200':
          description: Water history list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        intake_date: { type: string, format: date }
                        total_ml: { type: integer, example: 2500 }
                        amount_ml: { type: integer, example: 2500 }
                        entry_count: { type: integer, example: 3 }
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /me/water/{entryId}:
    delete:
      summary: Delete a logged water entry
      tags: [Water]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: entryId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Entry deleted
...
  /admin/settings:
    get:
      summary: Get application and system settings
      tags: [Admin, Settings]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: System and app settings
    put:
      summary: Update platform settings
      tags: [Admin, Settings]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Settings updated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...